Change Battery Icons to 2-Space Supported NF Characters
Some (mostly obsolete) NF character icons are not supported on the TMux
statusline and cause placement issues (like the previous "charged" icon)
while other (mostly brand-new) NF character iconsa re not supported
either in the Font (Caskadiya Kove NF) or the Terminal (Win Terminal).
Until we know for sure what the problem is, the best we can hope for is
to use characters compatible with both sides of the deal.