Temporary Fix: Update NeoVim's DevIcons Plugin for NF3
The de facto official GitHub repository for `nvim-web-devicons` at
`nvim-tree/nvim-web-devicons` is currently supporting only NF2, with a
pending PR for support to NF3. As my machines are (mostly) updated to
NF3, I'm using the forked (and patched) repo in said PR for the time
being until NF3 is fully supported in the main repo.