Port Global Vimscript Settings to Lua
Note that mappings are still in Vimscript - I have not yet made
up my mind as to whether I want to port them to Lua or not.

Additionally, I have not yet ported the plugin settings to Lua.
I'm still using `vim-plug` to manage plugins, and I'm not sure
if porting them to Lua is a good idea.