Enable Italics in TMux and Use for Neovim Comments
The original `default-terminal` settings for TMux was `screen-256color`,
which did not support italic fonts. The new setting is now
`tmux-256color`, which does support italics! This allows us to utilise
italic-rendering abilities within TMux as well, leading to...

The `Comment` highlight group is now shown in italics! The `gui` value
for the `Comment` highlight group is now set to `italic`, which renders
all `Comment`s (and inherited highlight groups) in italic fonts!