FEAT: Cross-platform Miniconda Script and `lsa`
Note that we are only considering user-local installations of Miniconda3
to either macOS (Darwin) or Linux at this time. Support for other
operating systems or installation directories will be added in the
future.

Also note that the `lsa` command for macOS systems require the
`coreutils` homebrew formula, which provides the GNU `ls` executable.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>