FEAT: Add Condition for VS Code's Neovim Extension
The Neovim extension for VS Code sources the same Neovim configurations
from this repository, which may cause undesireable side effects in the
VS Code environment due to compatibility issues.
As such, most configurations are now wrapped in a conditional statement
to check for the VS Code environment by default.
Note that while most configurations are currently unused in VS Code,
cursor selection options and virtual edit configurations are retained.

This commit is not yet tested on Linux or Mac environments.
However, as both Neovim and its embedded Lua support are both
platform-independent, stable results are expected on both environments
as well.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>