FEAT: Add Support for VS Code's Neovim Extension
The Neovim extension for VS Code sources the same configurations as
vanilla Neovim, which may cause severe errors in the VS Code
environment. As such, the configuration file is now wrapped in a
conditional statement to check for VS Code usage by default.
This will be sorted in future commits to replicate the current Neovim
setup in VS Code as much as possible.

This commit is not yet tested on Linux or Mac environments.
However, as both Neovim and its embedded Lua support are both
platform-independent, stable results are expected on both environments
as well.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>