FEAT: Remove Deprecated and Misleading Functions
The `sumatra` function was used to alias a specific PDF reader from
within WSL environments efficiently. As this is replaced with a more
stable and trustworthy functionality provided by the `wslu` package
(targeted specifically for WSL instances), this function will now be
deprecated.

The `tmux` function enforced usage of the `-2` flag when using TMux.
However, as both the function and its underlying command used the same
name, calling vanilla TMux from the fish shell was misleading.
This function is now removed to stop such misleading, bad behaviour.