CLEAN: Remove Deprecated and Misleading Functions
The `sumatra` function was used to alias a specific PDF reader from
within WSL environments efficiently.
As this is replaced with a more stable functionality provided by the
`wslu` package (specifically targeted for WSL), this function will now
be deprecated.

The `tmux` function enforced usage of the `-2` flag when using Tmux.
However, as both the function and its underlying command used the same
name, calling vanilla Tmux from the fish shell was misleading.
This function is now removed to stop such misleading behaviour.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>