FEAT: Right-flushed Prompt in PowerShell
Current (Mini/Ana)conda environment name has been moved to the
right-prompt area (called with `$Env:CONDA_PROMPT_MODIFIER`).
Current time and date is formatted and printed after the Conda
prompt modifier string.

Note that the current time and date string may be dependent on the MSYS2
`date` executable.
This is a Unix tool that may not be available on devices without MSYS2.
If so, this part of the prompt must be changed at a future time.