FEAT: Native Date Parsing in PowerShell Prompt
The `Get-Date` cmdlet for PowerShell is used instead of MSYS2's `date`
executable to parse and display current time and date on the prompt.
This ensures that the prompt can run without external dependencies.

Note that locale settings for `Get-Date` must be changed to "en-US"
for the intended formatting to correctly appear.
This is even more important for sessions whose locale have double-width
characters (such as Korean), as this can mess up cursor placement for
the right prompt and exceed the line width limit for the right-prompt.
The locale setting is therefore saved before the prompt renders,
changed to "en-US" during the prompt rendering, and restored upon
returning from the prompt-setter function.