FEAT: Update Pane Split for Neovim in Tmux
Removes horizontal splits when opening a new Neovim session,
as horizontal splits are too narrow in most cases for good use.

With tmux 3.4, the `-p` flag has been deprecated. The `-l` flag is now
used in its place, as well as an appending percentage symbol `%` to
denote that the new pane respects the current pane size.
This change is only imposed for the new tmux version 3.4,
which is tested by a conditional statement.