FEAT: Preliminary Options for `indent-blankline`
The `indent.tab_char` option is used to override hard-tab `lcs` configs
for leading tabs only.
Note that setting this as the same to `indent.char` may cause confusion
if tabs and spaces are used in the same file.
Highlights for all indentation is set equal to the highlight group
`NonText` at this time, to avoid confusion during the migration process.

The `whitespace.remove_blankline_trail` option is used to force
highlights for trailing whitespaces from the `indent-blankline` plugin.
This is unset at this time to allow other configurations to set
highlight groups, if necessary.

Other configuration options have not yet been set at this time.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>