FEAT: `indent-blankline` Indentation Visual Plugin
This commit introduces `indent-blankline` as a replacement to the
previously used `vim-indent-guides`.
This is both for a more Lua-centric plugin configuration ecosystem and
a more flexible indentation visual guide with many new functionalities.

The `indent.tab_char` option is used to override hard-tab `lcs` configs
for leading tabs only.
Note that setting this as the same to `indent.char` may cause confusion
if tabs and spaces are used in the same file.
Highlights for all indentation is set equal to the highlight group
`NonText` at this time, to avoid confusion during the migration process.

The `whitespace.remove_blankline_trail` option is used to force
highlights for trailing whitespaces from the `indent-blankline` plugin.
This is unset at this time to allow other configurations to set
highlight groups, if necessary.

Custom highlight groups are also added for indentation whitespaces.
The custom highlight groups `IndentOdd` and `IndentEven` are used for
odd and even levels of indentation, respectively.
Foreground-only highlight group `ScopeFG` is used to remove background
coloring from `indent-blanklink` scope highlights and preserve
highlighted indentation whitespace.
For the custom highlight groups to be recognised by the plugin on
startup, the `nvim_set_hl` API is called before the `setup()` function,
which is implemented in the `config` key.
The `opts` key is also used for better readability and code management
by dividing plugin-internal options (`opts`) and any possible external
or helper functions before or after plugin startup (`config`).

Values for the custom highlight groups are taken from the previous
configurations for the `vim-indent-guides` plugin.
These values are hard-coded, which may cause aesthetic issues when
the colorscheme is changed.

The current implementation for virtual text in Neovim does not support
visual mode highlight overlay over it.
As `indent-blankline.nvim` uses virtual text to draw the indent lines,
it is not possible to see the visual selection when `indent-blankline`
is enabled.
This commit provides a temporary workaround by disabling the indentation
guides entirely, for all buffers, when entering visual mode and enabling
them when exiting visual mode.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>