FEAT: Lua API to Set `EndOfBuffer` Highlight Group
The `nvim_get_hl` API is used to retrieve the baseline `Normal`
highlight group.
This is then used to create a table for a new highlight group,
`HideTilde`, which inherits the background color from `Normal`
and sets the foreground color to the background color.
The `HideTilde` highlight group is then set as the `EndOfBuffer`
highlight group, which is used to hide the buffer end tildes.

Unfortunately, the `EndOfBuffer` highlight group is currently
entirely overwritten with hard-coded values retrieved from the
`Normal` highlight group on startup.
This may cause the `EndOfBuffer` highlight group to be incorrect
if the `Normal` highlight group is modified (i.e. by a colorscheme)
after the `EndOfBuffer` highlight group is set.

Also changes the `init` key to `config` to suppress unexpected
behaviours that may occur when the plugin is loaded.
The `config` key is used to avoid calling a possible `setup()` function
embedded in the plugin.
As this is a Vimscript plugin, there is no `setup()` function to call,
but it is good practice to use the `config` key to avoid unexpected
`setup()` function calls.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>