FEAT: Custom Highlight Groups for Indentation
This commit adds custom highlight groups for indentation whitespaces
in the `indent-blankline.nvim` plugin.
The custom highlight groups are `IndentOdd` and `IndentEven`, used for
odd and even levels of indentation, respectively.
At this time, only background colors for both GUI and TUI are set.
Foreground colors will be added if necessary at a later commit.

Also changes the `opts` key to `config` to call additional functions
before the actual `setup()` function is called.
This forces the `nvim_set_hl` API to be called before the `setup()`
function is called, which is necessary for the custom highlight groups
to be recognised by the plugin on startup.

Values for the custom highlight groups are taken from the previous
configurations for the `vim-indent-guides` plugin, which is no longer
used here.
These values are hard-coded, which may cause aesthetic issues when
the colorscheme is changed.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>