FEAT: Toggle `indent-blankline` in Visual Mode
The current implementation for virtual text in Neovim does not support
visual mode highlight overlay over it.
As `indent-blankline.nvim` uses virtual text to draw the indent lines,
it is not possible to see the visual selection when `indent-blankline`
is enabled.
This commit provides a temporary workaround by disabling the indentation
guides entirely, for all buffers, when entering visual mode
and enabling them when exiting visual mode.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>