REFACTOR: Move Plugin Configs to Options Block
The `config` block is used to sequentially call multiple functions
related to the plugin, part of which is the `setup()` function provided
by the plugin itself.
As options for the plugins can be stored in a separate `opts` block
and read in the `config` block, this commit splits the plugin-specific
option table from other configuration functions for better readability.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>