REFACTOR: Replace Legacy Vimscript Configs to Lua
Refactors the legacy Vimscript-written configuration snippets to Lua
for efficiency and readability.
Most of the ported configurations are mappings, which will be further
refactored in future commits.

Filetype detection for `*.h` files is moved to the `filetype.vim` file,
stored in the Neovim configuration top directory.

Note that the `Comment` highlight group is modified at the end of the
configuration file, after the Lazy.nvim setup, which includes a plugin
for colorscheme management.
This ensures that the `Comment` highlight group is modified after the
colorscheme is loaded, which is necessary for changes to take effect.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>