FEAT: Use MSYS2 Binaries in PowerShell Functions
As a start, the `ls` command, previously used as an alias for the
`Get-ChildItem` PowerShell commandlet, is now removed in favour of the
Linux `ls` command, implemented in MSYS2 as `ls.exe`.
As this program is capable of receiving various flags as arguments,
an additional `lsa` function is defined to easily use the `ls.exe`
executable with frequently-used flags pre-enabled.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>