FEAT: Remove RVM Dependency for Ruby in Fish Shell
RVM's extensive manipulation of shell environments, as well as its
side effects of "hijacking" commands and functions under-the-hood,
makes it difficult for me as a maintainer of my system to clearly
understand what each aspect of my configurations are responsible for.
As I do not need a version manager for Ruby at this time,
the RVM dependency will be removed at this time.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>