FEAT: `fisher` as a Plugin Manager for Fish Shell
Currently installed plugins are:
- `jorgebucaran/fisher`, the base `fisher` plugin file, AND
- `edc/bass`, a bash script execution helper for env-vars and v-envs

Note that `fisher` intrudes on Git, which allows all systems at a
specific commit to share all plugins and versions.
This may have unintended side-effects, upon which the `$fisher_path`
installation path will be specified and removed from the repository.
In such a case, links to the fish-shell config directory may need to be
established for the plguins to function correctly.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>