REFACTOR: Move Treesitter Config Options to `opts`
As the function previously defined in `config` was only used to call
the main setup function in `nvim-treesitter.configs`, this is split up
into options `main` and `opts`, with `main` referring to the location of
the `setup` function to call and `opts` providing options passed to said
`setup` function.

Also removes the `commit` option to track the latest changes provided
for Treesitter.
This is combined with the decision to ignore installation for the
`latex` parser, preventing installation from source for all currently
known and used languages, even for environments without the
`tree-sitter` CLI.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>