FEAT: `telescope.nvim` Versatile Fuzzy-finder
This commit introduces the `telescope.nvim` versatile fuzzy-finder.
To enable fuzzy-finding, a dependency plugin `telescope-fzf-native.nvim`
must also be installed.
Normally on Unix-based machines, this plugin would be installed via a
pre-written Makefile by running `make`, but as the raw `make` command
is not supported on Windows systems, an extensive conditional block is
used to ensure smooth, automated installation.
Window systems depend on CMake and consecutively execute multiple steps
to build the final library output.

The current implementation supports Visual Studio 2022 and MinGW as
possible generators for CMake on Windows.
Visual Studio 2022 is tried first, and if it fails, a `git clean -Xfd`
command is passed to the shell to remove previous failed build outputs,
after which MinGW is tried as an alternative.
Should both generators fail, the shell's error message is passed into
Neovim, followed by a short addendum from the build function.

This new function also supplies a local function `throw_err` that
receives the output object of a system command and returns the error
message as-is if the object's `code` is nonzero.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>