FEAT: LSP Configurations with Added Completions
The `mason` plugin manager specialises in LSP, DAP, linter and
formatter plugins for Neovim.
This plugin manager is bootstrapped as a plugin alongside LSP-focused
plugins such as `mason-lspconfig` and `nvim-lspconfig` to integrate LSP
functionalities into Neovim runtimes as smooth as possible.

`nvim-telescope` and `nvim-cmp` are integrated with LSP functionalities
for an even smoother experience - `nvim-telescope` is used to check for
diagnostics provided by the LSP, while `nvim-cmp` is further extended
for completions provided by the LSP.
In particular, the LSP support in `nvim-cmp` is followed up by the usage
of `cmp-nvim-lsp`, an additional plugin specifically for LSP
integrations with `nvim-cmp`.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>