FIX: Build Function for Telescope FZF Module
As the raw `make` command is not supported on Windows systems,
this commit replaces the previous string to a function that executes
different system commands based on the detected OS.
While Unix-like operating systems still use the `make` command,
it is now explicitly passed to the system as a blocking command.
Window systems depend on CMake and consecutively execute multiple steps
to build the final library output.

The current implementation supports Visual Studio 2022 and MinGW as
possible generators for CMake on Windows.
Visual Studio 2022 is tried first, and if it fails, a `git clean -Xfd`
command is passed to the shell to remove previous failed build outputs,
after which MinGW is tried as an alternative.
Should both generators fail, the shell's error message is passed into
Neovim, followed by a short addendum from the build function.

This new function also supplies a local function `throw_err` that
receives the output object of a system command and returns the error
message as-is if the object's `code` is nonzero.

With the addition of an extesive build function to the FZF module,
the plugin spec is extracted from the `dependencies` table where it was
before, into its own table.
This file now provides two side-by-side tables to the plugin manager
for this module and the original `telescope.nvim` plugin.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>