FEAT: Support Single-user Nix for Login Shells
Requires the `nix-user-chroot` executable in path, preferably the
`.local/bin` directory.
This executable effectively adds the `/nix` directory at filesystem root
by "chroot"ing into a user namespace while bind-mounting the rest of the
filesystem as well.
For systems without the `nix-user-chroot` executable available, the
original `tmux -2` command is executed.

Find `nix-user-chroot` at github.com:nix-community/nix-user-chroot.
The executable can be downloaded as a binary file from the releases
section and used without Rust installed on the system.

Note that creating user namespaces must be enabled for the executable
to work; this requires a Linux kernel of version 3.8 or higher,
with the necessary kernel abilities enabled.
Notably, the executable may not work on RHEL or CentOS 7.x systems that
do not have this functionality enabled by default.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>