FEAT: Use and Favour `eza` for `ls` Overriding
While the MSYS2 version of `ls` is usable in native Windows, it does not
support Unicode characters, which cause problems when displaying Korean
characters.
This problem is mitigated with the usage and favouring of `eza`,
a cross-platform tool that runs natively on all major operating systems
while also supporting Unicode in native Windows.
The introduction of `eza` also brings more functionality to the shell,
as it supports multiple different colors and icons according to
individual filetypes, while supporting most of the options previously
used with `ls`.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>