FEAT: Print Last Exit Code in PowerShell Prompt
Records the global last exit code at the start of the prompt function,
which is used to produce either a success or failure icon.
The failure icon also shows the non-zero exit code.

Note that the global last exit code may not be overwritten for some
cmdlets; notably, the `clear` cmdlet does not reset this code.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>