REFACTOR: Import `nix-user-chroot` Fish Script
While the automatic login to a Nix-package environment is only enabled
in the `nix-chroot` branch, we may want to try `nix-user-chroot` with
our main branch for testing purposes.
This commit imports the corresponding initialiser script that allows the
fish shell to find Nix executables once logged in with the
`nix-user-chroot` executale.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>