FEAT: New Function `tree` for Tree View with `eza`
The fish-shell function `tree` is used to display a tree-view of the
current directory with a default depth of 1, while also providing a
long-view with some default settings.
This function is dependent on the `eza` executable, which is taken into
account by encasing the function definition in a conditional statement.

Future commits may introduce similar functionality to PowerShell.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>