FEAT: Add Wrapper Function for Work Server Scripts
`ic`-scripts written in Bash are unusable in Fish, as Fish does not
allow Bash scripts to alter path variables.
This function is a workaround to provide a new Fish shell instance from
Bash, where the scripts are run prior to entering the Fish shell.
This allows the subsequent Fish shell instance to inherit all
environmental variables, including those from the script, to be used in
itself.

Note that this function is to be used only on specific servers that use
these custom-drafted scripts.
I do not take any responibility for usage and consequences of this
function outside its intended, personal scope of use.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>