FEAT: Env Variables for Mounted PDK `home` Drives
This snippet traverses through results from the `mount` command to
search for other PDK server home drives mounted on the server.
Accessible home drives are then assigned global env variables to easily
access `$HOME` directories for this user on other servers.
These env variables persist for the remainder of this session and are
updated for each instance.

Metadata such as the IP address for each recognised server are printed
on session startup, alongside the `$fish_greeting` message.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>