STYLE: Update PowerShell's `tree` Function
This commit follows the previous commit `79ea3eb` to remove the
default level option in the `tree` function, this time in PowerShell.

Having a default level removes the possibility of displaying all file
trees open, i.e. without a level option.
This commit removes the default level option embedded in the function
call, which allows for a full tree-view.
Levels can be applied with the `-L` option.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>