FIX: Use Reliable Indicator for TMUX Environment
The `tmux showenv` command produces a successful exit code inside a TMUX
environment and a failed exit code outside.
This is a better indicator than using environmental variables, as these
can last outside of the current TMUX environment and bleed in to other
shells.

Also refactors the conditional statement according to the new condition
that is truthy when inside TMUX.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>