FIX: Remove `conda`'s default Right-side Prompt
Fish shell installed by `conda` have a specific snippet loaded after the
user snippets specifically for the `conda` environment, in which is
included a `__conda_add_prompt` function to add the right-side prompt.
As this overwrites the user-defined `__conda_add_prompt` function
intended to nullify it, this function is moved to the user's
`config.fish` to ensure that it works the intended way.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>