FEAT: Move Fish Function `ic` and Add Completions
Function `ic` is now implemented in its own file under the
`./fish/functions` directory and is able to be called in all hosts.

The `argparse` builtin command is used to easily parse flags, with the
`-h/--help` flag implemented as a test case.
Note that the help flag prints hardcoded strings instead of calling a
pager to show a dedicated manpage; this may be changed in future commits
if applicable.

This commit also introduces a completion snippet for the `ic` function,
which preemptively searches for available subcommands and supplies
completions.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>