FEAT(fish): Args for Help and User-wise VNC Report
This adds a preliminary help message (as `-h/--help`) and a user-wise
reporting option (`-u/--user[=?]`) for the `vnclist` function.

The user-wise reporting option (`-u/--user[=?]`) can receive up to one
optional argument.
If an optional argument is given, the function prints all ports used by
the provided user.
If an optional argument is NOT given (with only `-u/--user` as input),
the function prints all currently inhabiting users and their respective
ports.

Output formats may change in future updates to this function.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>