REFACTOR: Drop Default Menu and Keybinding Configs
As the default configurations are applied as a fallback,
the addition of these files to the repository serve as a starting point
for personalised configurations.
As menus and keybinding are not yet planned for extensive customisation,
these settings will be dropped and called upon in fallback at this time.

Note that default configurations and environmental settings can easily
be accessed with:
    config [nu/env] --default

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>