FEAT: Aliases and Functions from External Files
The `./aliases.nu` and `./functions.nu` files are separated from the
main configurations file for better organisation.
As there is no automatic method of importing snippets from other files,
this commit manually `source`s the necessary files at the end of
configuration.

The `./aliases.nu` file notably comes with two new aliases `lsa` and
`tree`, which are based on the `eza` binary and used to list files.
The `./functions.nu` file is empty at this time; this file will mainly
be used for functions that are capable of using Nushell-oriented data
types.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>