FEAT: Basic In-house Git Prompt
A helper function `__is_git_dir` is defined to detect if the current
directory is a git directory, providing an easy one-shot solution to
detecting and printing repo-related information.
This function is then called in the command prompt function.

This helper function is still a work in progress; future commits may
introduce changes to further enhance its usability.
The ultimate goal here is to mimic and supersede the git-prompt provided
by default in the fish shell.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>