FEAT: Basic Implementation of `lscd` Functionality
This commit adds a basic implementation of the `lscd` function used and
defined in my other shell profiles.
It depends on the built-in `cd` command for parsing, retaining most of
its functionality, including the usage of `-` to move back to the
previous directory.

Additional functionalities and safety measures will be implemented in
future commits.
Notably, native autocompletion is currently a work-in-progress.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>