FEAT: Configuration for Verilog LSP (`verible`)
The Verible LSP for (System)Verilog is assigned a basic configuration,
with focus on finding the Verible linting rules file and basing both the
LSP root directory and starting command on it.
Other functions such as auto-completion and inter-file linking may not
yet be fully configured; future commits will debug such issues, if any.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>