REFACTOR: Lua Function to Save Without Whitespaces
A custom function `save_without_trailing_whitespaces()` is newly
introduced to handle the job of searching for trailing whitespaces and
replacing them without altering the search history, as well as updating
the file open in the current buffer only if it has not been changed.
Additionally, this new function keeps the current cursor position during
its run.

<C-s> mappings are now implemented in Lua near other Lua keymaps,
with the same `save_without_trailing_whitespaces()` function called
in both normal and insert modes.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>