FEAT: New Fish Function `pdk` for More Flexibility
A new function `pdk` is implemented to expand `source`-able PDK scripts
from eligible servers and hosts, by receiving available resources from
the `/opt/mi-env/bin/` directory.
This function is defined in its own file under the `./fish/functions`
directory and is able to be called in all hosts, while the related
completions script is located under `./fish/completions`.

Most of the function structure and abilities are directly taken from the
preceding `ic` function, which will be deprecated once this function and
its completions are successfully tested over time.
Other functionalities and/or hardened security measures may be
implemented for this function and its completion in future commits.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>