FEAT: Filetype Recognition for Tcl Files
Similarly to commit 5e0569c, Tcl filetype recognition is added for
Synopsys configuration files and Innovus command logs, each based on
the filename pattern `**.synopsys_dc.setup*` and `**/innvous.cmd*`.
Note that, while the Innovus command logs are not Tcl scripts, they are
records of Tcl commands that can be used in place of a script, and as
such would benefit from this configuration.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>