FEAT: Interactive `rm` Blockage Function
As a habitual `rm` can cause irreversible changes, it would seem best to
refrain from using it in interactive settings.
As such, a new function `rm` is defined to override the original `rm`
command in interactive environments to print a warning statement.
Use an XDG-compliant trash utility such as `gio trash` or `trash-cli`
instead to minimise possible damages to a system.

This change only applies to interactive sessions to avoid potential
problems when scripts are executed with configurations applied.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>