FEAT: Configuration for Built-in Nushell LSP
Nushell's built-in LSP can be set up in Neovim using default
configurations supplied by the `nvim-lspconfig` plugin.
As this is an external LSP not managed within Neovim (through Mason),
an executable Nushell binary is necessary for this to work.
Attempting to call the Nushell LSP on systems without Nushell have not
been tested at this time.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>