FEAT: Additional Windows Configuration for GnuPG
As Git for Windows is based on MSYS2 for easier distribution of Git's
Unix-based codebase, the default GnuPG application used on Windows is
also dependent on MSYS2.
As installing a full MSYS2 distribution solely for GnuPG integration
with Git for Windows is cumbersome, a native GnuPG application is
installed and used as a replacement, with the `gpg.program` option
additionally set to account for this change.
This option is intended to be applied only on Windows systems;
as such, a separate configuration file is introduced to house
Windows-specific options, to be (sym)linked to the `%USERPROFILE%`
directory as another source of "global" configuration.
The README documentation has been updated to include this additional
setup during installation.

However, the fact that this change causes minor compatibility issues
between Git and GnuPG runtimes, as well as native GnuPG on Windows being
less stable than its Unix-based counterpart, causes issues when first
invoking GnuPG for commit signing.
Noted problems range from noticeable lag during commit signing to errors
during commit.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>