NEW: Nushell Configurations Branch
As Nushell uses `nu` as its CLI command, its corresponding subtree
branch for DOTS will also be named `subtree-nu` as well.
However, configurations for Nushell must be placed in the
`$XDG_CONFIG_HOME/nushell` directory, and as such the subdirectory in
which this subtree will be merged into will also be named `nushell`.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>