NEW: PowerShell Configurations Branch
As PowerShell uses `pwsh` as its CLI command, its corresponding subtree
branch for DOTS will also be named `subtree-pwsh` as well.
However, configurations for PowerShell must be placed in the
`$env:USERPROFILE\Documents\PowerShell` directory, and as such the
subdirectory in which this subtree will be merged into will also be
named `PowerShell`.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>