FEAT: Function Declarations for PowerShell REPL
Functions previously defined in other shells such as `lsa`, `lscd` and
`tree` are defined with native cmdlets and external commands.
The default alias `ls` for the PowerShell cmdlet `Get-ChildItem` is
removed in favour of the `eza` executable, with additional options for
further personalisation.

While most dependencies are PowerShell internals or native to Windows,
some fallback options may depend on the MSYS2 distribution, which can
cause compatibility issues if the unavailable fallbacks are necessary.
Caution is advised.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>