FEAT: `ic*`-related PDK Server Snippets
This commit introduces the `pdk.fish` snippet activated for servers for
which the hostname starts with "pdk".
This conditional statement effectively ensures that the snippet is
enabled for PDK servers only.

The `ic` function searches for executable commands that set global
environmental variables for versioned tools such as the Cadence suite
and Matlab.
The `mount`-traversing snippet below sets global environmental variables
for other PDK servers' home drives to be easily accessible for the
current user.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>