FIX: Find Failed Commit Message from Worktrees
Following up on commit aeb6f6e in reusing previous failed commits,
where it is noted that commit messages cannot be reliably found for
worktrees with the top-directory `.git` being a file instead of the
normal metadata directory, this commit updates the previous alias to be
able to find the correct commit message file with the help of the shell.
Extra flags and arguments are passed into the command with the use of
`$@`, intentionally left (slightly) unprotected to favour flexibility.

Note that this alias uses the `sh` command under the hood to enable the
use of command substitution.
As such, `sh` must be available in $PATH when this alias is invoked and
be a shell conformant to POSIX standards.
For Unix-based platforms this will mostly not be a problem as `sh` is
most likely to be the system Bash shell.
For Windows platforms the default distribution "Git for Windows" comes
with a bundled and executable Bash shell, but this must be available in
the shell on which Git is used and this alias is invoked.

Future commits may look into alternative ways to achieve this
functionality, preferably through Git internals and more native methods.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>