FIX: Suppress `read` in `rm` Function Evaluation
As the Fish shell autocompletion engine partially evaluates functions
for tab completion after startup, the internal `read` function is
executed and effectively hijacks the command line when TYPING the
`rm` function for the first time in any Fish shell instance.
As such, an additional condition is placed to guard this part of the
function in such cases using the "current command".
This value is normally "rm" when interactively run but is "fish" when
evaluated by the Fish shell autocompletion engine and therefore can be
compared against the "function name" (which is always "rm") to confirm
the context of the function call.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>