CHORE: Banner "Uptime" Format Update
Adds an "and" in place of the final joining comma for a more natural
look and feel.

Also refactors the plurality check for each element of the time-unit
where an additional check `$in.num != 0` is unnecessary, as in a record
converted from a duration, all values are positive integers;
the plurality check only needs to check if each value is `1` or higher.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>